# contract/openapi.yaml
openapi: 3.1.0
info:
  title: AI 健身助手 P0 API
  version: 1.0.0
  description: >-
    微信首发但渠道无关的业务契约。关键训练数字由版本化确定性规则产生，
    AI 仅提供候选、排序和解释；用户归属始终由认证主体确定。
servers:
  - url: /
    description: local/test relative base
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/exercises:
    get:
      tags: [Content]
      operationId: listExercises
      parameters:
        - {name: equipmentType, in: query, required: false, schema: {type: string}}
        - {name: movementPattern, in: query, required: false, schema: {type: string}}
        - {name: muscleGroup, in: query, required: false, schema: {type: string}}
      responses:
        '200':
          description: 当前用户器械可支持且符合当前环境发布级别的动作
          content:
            application/json:
              schema: {$ref: './schemas/content.yaml#/components/schemas/ExerciseListResponse'}
        default: {$ref: '#/components/responses/DefaultError'}
  /api/v1/exercises/{id}:
    get:
      tags: [Content]
      operationId: getExercise
      parameters:
        - {$ref: '#/components/parameters/IdPath'}
      responses:
        '200':
          description: 动作详情、通俗解释、图片回退引用与合法替代动作
          content:
            application/json:
              schema: {$ref: './schemas/content.yaml#/components/schemas/ExerciseDetailResponse'}
        '404': {$ref: '#/components/responses/NotFound'}
        default: {$ref: '#/components/responses/DefaultError'}
  /api/v1/exercises/{sourceCode}/replacements:
    get:
      tags: [Content, Workouts]
      operationId: listExerciseReplacements
      description: 仅返回当前用户器械可用、未排除且审核状态合格，并且动作模式、难度、主要肌群、负重模式与处方转换兼容的替代动作；不修改计划或训练快照。
      parameters:
        - {name: sourceCode, in: path, required: true, schema: {type: string, pattern: '^[A-Z0-9_]+$'}}
      responses:
        '200':
          description: 1 至 4 个通用安全替代候选；无合法候选时宁缺毋滥
          content:
            application/json:
              schema: {$ref: './schemas/content.yaml#/components/schemas/ExerciseReplacementResponse'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/DefaultError'}
        default: {$ref: '#/components/responses/DefaultError'}
  /api/v1/workout-sessions/{sessionId}/exercises/{snapshotId}/replacements:
    get:
      tags: [Workouts]
      operationId: listWorkoutExerciseReplacements
      description: 绑定当前用户的训练会话和快照，排除本次训练已有动作后返回安全替代候选。
      parameters:
        - {name: sessionId, in: path, required: true, schema: {type: string, format: uuid}}
        - {name: snapshotId, in: path, required: true, schema: {type: string, format: uuid}}
      responses:
        '200':
          description: 1 至 4 个不与本次训练其他动作重复的安全候选
          content:
            application/json:
              schema: {$ref: './schemas/content.yaml#/components/schemas/ExerciseReplacementResponse'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/DefaultError'}
        default: {$ref: '#/components/responses/DefaultError'}
...
  /api/v1/plans/{planId}/exercise-replacements:
    get:
      tags: [Plans]
      operationId: listPlanExerciseReplacements
      description: 返回指定计划日内原动作的已审核严格等价替换；候选保持原组数、次数和休息，仅在负重模式变化时要求重新校准。
      parameters:
        - {name: planId, in: path, required: true, schema: {type: string, format: uuid}}
        - {name: dayCode, in: query, required: true, schema: {type: string, minLength: 1, maxLength: 64, pattern: '^[A-Z0-9_]+$'}}
        - {name: sourceExerciseCode, in: query, required: true, schema: {type: string, minLength: 1, maxLength: 64, pattern: '^[A-Z0-9_]+$'}}
      responses:
        '200':
          description: 同动作模式、主要肌群和难度且符合当前器械与排除偏好的替换候选；无合格项时 items 为空。
          content:
            application/json:
              schema: {$ref: './schemas/plan.yaml#/components/schemas/PlanExerciseReplacementOptionListResponse'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '404': {$ref: '#/components/responses/NotFound'}
        default: {$ref: '#/components/responses/DefaultError'}
...
  /api/v1/workout-sessions/{id}/exercises/{exerciseId}:
    put:
      tags: [Workouts]
      operationId: updateWorkoutSessionExercise
      parameters:
        - {$ref: '#/components/parameters/IdPath'}
        - {$ref: '#/components/parameters/ExerciseIdPath'}
      requestBody: {$ref: '#/components/requestBodies/WorkoutExerciseUpdate'}
      responses:
        '200':
          description: 仅当前训练生效的替换覆盖层；原计划版本和原始训练快照保持不变
          content:
            application/json:
              schema: {$ref: './schemas/workout.yaml#/components/schemas/WorkoutSessionResponse'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '409':
          description: 版本冲突，或当前训练已无不重复且安全的替代动作
          content:
            application/json:
              schema: {$ref: './schemas/workout.yaml#/components/schemas/WorkoutExerciseUpdateConflictResponse'}
        default: {$ref: '#/components/responses/DefaultError'}
...
  /api/v1/progress/exercises/{exerciseId}:
    get:
      tags: [Progression]
      operationId: getExerciseProgress
      parameters:
        - {$ref: '#/components/parameters/ExerciseIdPath'}
      responses:
        '200':
          description: 有效正式组趋势
          content:
            application/json:
              schema: {$ref: './schemas/progression.yaml#/components/schemas/ExerciseTrendResponse'}
        default: {$ref: '#/components/responses/DefaultError'}
  /api/v1/progression-recommendations:
    get:
      tags: [Progression]
      operationId: listProgressionRecommendations
      parameters:
        - {name: status, in: query, required: false, schema: {$ref: './schemas/progression.yaml#/components/schemas/RecommendationStatus'}}
        - {name: cursor, in: query, required: false, description: Base64URL 编码的稳定游标, schema: {type: string, maxLength: 256}}
        - {name: limit, in: query, required: false, schema: {type: integer, minimum: 1, maximum: 100, default: 20}}
      responses:
        '200':
          description: 有界建议列表，按 createdAt、id 倒序
          headers:
            X-Has-More:
              required: true
              schema: {type: boolean}
            X-Next-Cursor:
              required: false
              schema: {type: string}
          content:
            application/json:
              schema: {$ref: './schemas/progression.yaml#/components/schemas/ProgressionRecommendationListResponse'}
        default: {$ref: '#/components/responses/DefaultError'}
  /api/v1/progression-recommendations/{id}/apply:
    post:
      tags: [Progression]
      operationId: applyProgressionRecommendation
      description: USER_LOCKED 重量只允许提示；应用前必须显式确认并创建新计划版本。
      parameters:
        - {$ref: '#/components/parameters/IdPath'}
        - {$ref: './schemas/common.yaml#/components/parameters/IdempotencyKey'}
      requestBody: {$ref: '#/components/requestBodies/RecommendationApply'}
      responses:
        '200':
          description: 已采纳建议
          content:
            application/json:
              schema: {$ref: './schemas/progression.yaml#/components/schemas/ProgressionRecommendationResponse'}
        '409': {$ref: '#/components/responses/VersionConflict'}
        default: {$ref: '#/components/responses/DefaultError'}
  /api/v1/progression-recommendations/{id}/dismiss:
    post:
      tags: [Progression]
      operationId: dismissProgressionRecommendation
      parameters:
        - {$ref: '#/components/parameters/IdPath'}
      requestBody: {$ref: '#/components/requestBodies/RecommendationDismiss'}
      responses:
        '200':
          description: 已忽略建议
          content:
            application/json:
              schema: {$ref: './schemas/progression.yaml#/components/schemas/ProgressionRecommendationResponse'}
        default: {$ref: '#/components/responses/DefaultError'}
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
  parameters:
    IdPath: {name: id, in: path, required: true, schema: {type: string}}
...
    ExerciseIdPath: {name: exerciseId, in: path, required: true, schema: {type: string}}
...
  requestBodies:
...
    WorkoutExerciseUpdate: {required: true, content: {application/json: {schema: {$ref: './schemas/workout.yaml#/components/schemas/UpdateSessionExerciseRequest'}}}}
...
    RecommendationApply: {required: true, content: {application/json: {schema: {$ref: './schemas/progression.yaml#/components/schemas/ApplyRecommendationRequest'}}}}
    RecommendationDismiss: {required: true, content: {application/json: {schema: {$ref: './schemas/progression.yaml#/components/schemas/DismissRecommendationRequest'}}}}
...
  responses:
...
    BadRequest: {$ref: './schemas/common.yaml#/components/responses/BadRequest'}
    Unauthorized: {$ref: './schemas/common.yaml#/components/responses/Unauthorized'}
    NotFound: {$ref: './schemas/common.yaml#/components/responses/NotFound'}
    VersionConflict: {$ref: './schemas/common.yaml#/components/responses/VersionConflict'}
...
    DefaultError: {$ref: './schemas/common.yaml#/components/responses/DefaultError'}